complete template for hero page
