+++ .prettierrc
@@ -1,17 +1,17 @@
 {
-  "trailingComma": "es5",
-  "tabWidth": 4,
-  "semi": true,
-  "singleQuote": false,
-  "overrides": [
-    {
-      "files": "*.prisma",
-      "options": {
-        "parser": "prisma",
-        "semi": true,
-        "singleQuote": true,
-        "trailingComma": "all"
-      }
-    }
-  ]
+    "trailingComma": "es5",
+    "tabWidth": 4,
+    "semi": true,
+    "singleQuote": false,
+    "overrides": [
+        {
+            "files": "*.prisma",
+            "options": {
+                "parser": "prisma",
+                "semi": true,
+                "singleQuote": true,
+                "trailingComma": "all"
+            }
+        }
+    ]
 }
